leaving one's old life for love ===============================
:: Description
Someone struggles with the life choice of whether to leave everything else behind to follow a lover.

:: Parents
leaving one's old life behind

money can't buy love
====================

:: Description
This age old adage is demonstrated, refuted, or otherwise discussed.

:: Parents
life lesson

coping with a tough customer
============================

:: Description
The proprietor of a business struggles to maintain cordial relations with a client who refuses to be content.

:: Parents
coping with problematic people

war due to diplomatic miscalculation
====================================

:: Description
Neither faction in a diplomatic negotiation really wants war but fluke and miscalculation threaten to pull them into a military confrontation anyway.

:: Parents
diplomacy

imperialism vs. modus vivendi
=============================

:: Description
A faction which espouses a policy of expansion through force is contrasted with a faction that seeks a way to co-exist amicably with their neighbors.

:: Parents
social dichotomy